Move address picker to component sub folder
